Re Organized Packages in instagrammers' story
presenter and model packages goes into view model (because they are part of the MVP).
