+++ app/src/main/java/com/losextraditables/bu/instagrammers/view/activity/InstagrammersListActivity.java
@@ -9,8 +9,8 @@ import com.karumi.rosie.view.Presenter;
 import com.losextraditables.bu.R;
 import com.losextraditables.bu.base.view.activity.BuActivity;
 import com.losextraditables.bu.instagrammers.InstagrammersListModule;
-import com.losextraditables.bu.instagrammers.model.InstagrammerModel;
-import com.losextraditables.bu.instagrammers.presenter.InstagrammersListPresenter;
+import com.losextraditables.bu.instagrammers.view.model.InstagrammerModel;
+import com.losextraditables.bu.instagrammers.view.presenter.InstagrammersListPresenter;
 import com.losextraditables.bu.instagrammers.view.adapter.InstagrammersAdapter;
 import com.losextraditables.bu.login.activity.LoginActivity;
